requirements.txt.txt
streamlit
pandas
openpyxl
google-generativeai
numpy